+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Changed
+- add `{{loading-indicator}}` default message string
 
 ## [0.5.4]
 - Preview button now takes user to the selected item [Link](https://esriarlington.tpondemand.com/entity/83079)